# backend/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: CarbonLedger API
  version: 1.0.0
  description: "Verified carbon credits. Permanent retirement. Full provenance.\n\n\
    ## Authentication\n- **JWT** (Bearer) \u2014 obtain via `POST /api/v1/auth/verify`.\n\
    - **API Key** (`X-Api-Key`) \u2014 for public API gateway endpoints.\n\n## Versioning\n\
    All routes are served under `/api/v1/`.\n\n## CarbonError codes\n| Code | Name\
    \ |\n|------|------|\n| 1 | ProjectNotFound |\n| 2 | ProjectNotVerified |\n| 3\
    \ | ProjectSuspended |\n| 4 | InsufficientCredits |\n| 5 | AlreadyRetired |\n\
    | 6 | SerialNumberConflict |\n| 7 | UnauthorizedVerifier |\n| 8 | UnauthorizedOracle\
    \ |\n| 9 | InvalidVintageYear |\n| 10 | ListingNotFound |\n| 11 | InsufficientLiquidity\
    \ |\n| 12 | PriceNotSet |\n| 13 | MonitoringDataStale |\n| 14 | DoubleCountingDetected\
    \ |\n| 15 | RetirementIrreversible |\n| 16 | ZeroAmountNotAllowed |\n| 17 | ProjectAlreadyExists\
    \ |\n| 18 | InvalidSerialRange |"
  contact:
    name: CarbonLedger
    url: https://carbonledger.io
servers:
- url: https://api.carbonledger.io
  description: Production
- url: http://localhost:3001
  description: Local development
...
paths:
...
  /api/v1/public/projects:
    get:
      tags:
      - projects
      summary: List public approved projects
      operationId: projects_get_list
      responses:
        '200': *id002
        '400': *id003
        '401': *id004
        '404': *id005
        '500': *id006
      security: *id001
...
  /api/v1/credits/batch/{id}:
    get:
      tags:
      - credits
      summary: Get credit batch
      operationId: credits_get_get
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreditBatch'
      security: *id001
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /api/v1/credits/retirement/{id}:
    get:
      tags:
      - credits
      summary: Get retirement record
      operationId: credits_get_get
      responses:
        '200': *id002
        '400': *id003
        '401': *id004
        '404': *id005
        '500': *id006
      security: *id001
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /api/v1/credits/lookup/{serial}:
    get:
      tags:
      - credits
      summary: Lookup credit by serial number
      operationId: credits_get_lookup
      responses:
        '200': *id002
        '400': *id003
        '401': *id004
        '404': *id005
        '500': *id006
      security: *id001
      parameters:
      - name: serial
        in: path
        required: true
        schema:
          type: string
  /api/v1/credits/provenance/{serial}:
    get:
      tags:
      - credits
      summary: Get credit provenance trail
      operationId: credits_get_get
      responses:
        '200': *id002
        '400': *id003
        '401': *id004
        '404': *id005
        '500': *id006
      security: *id001
      parameters:
      - name: serial
        in: path
        required: true
        schema:
          type: string
  /api/v1/credits/mint:
    post:
      tags:
      - credits
      summary: Mint credits
      operationId: credits_post_mint
      responses:
        '200': *id002
        '400': *id003
        '401': *id004
        '404': *id005
        '500': *id006
      security: *id009
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MintCreditsDto'
            example:
              projectId: proj_123
              amount: 100
              vintageYear: 2024
...
  /api/v1/credits/retire:
    post:
      tags:
      - credits
      summary: Retire credits
      operationId: credits_post_retire
      responses:
        '200': *id002
        '400': *id003
        '401': *id004
        '404': *id005
        '500': *id006
      security: *id007
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetireCreditsDto'
            example: &id010
              batchId: batch_abc
              amount: 5
              beneficiary: Acme Corp
              purpose: offset
...
  /api/v1/oracle/status/{projectId}:
    get:
      tags:
      - oracle
      summary: Oracle status for project
      operationId: oracle_get_oracle
      responses:
        '200': *id002
        '400': *id003
        '401': *id004
        '404': *id005
        '500': *id006
      security: *id001
      parameters:
      - name: projectId
        in: path
        required: true
        schema:
          type: string
...
  /api/v1/oracle/ingest/monitoring:
    post:
      tags:
      - oracle
      summary: Ingest monitoring data (oracle)
      operationId: oracle_post_ingest
      responses:
        '200': *id002
        '400': *id003
        '401': *id004
        '404': *id005
        '500': *id006
      security: &id011
      - X-Api-Key: []
  /api/v1/oracle/ingest/batch-monitoring:
    post:
      tags:
      - oracle
      summary: Ingest batch monitoring (oracle)
      operationId: oracle_post_ingest
      responses:
        '200': *id002
        '400': *id003
        '401': *id004
        '404': *id005
        '500': *id006
      security: *id011
  /api/v1/oracle/ingest/price:
    post:
      tags:
      - oracle
      summary: Ingest price (oracle)
      operationId: oracle_post_ingest
      responses:
        '200': *id002
        '400': *id003
        '401': *id004
        '404': *id005
        '500': *id006
      security: *id011
  /api/v1/oracle/ingest/batch-price:
    post:
      tags:
      - oracle
      summary: Ingest batch price (oracle)
      operationId: oracle_post_ingest
      responses:
        '200': *id002
        '400': *id003
        '401': *id004
        '404': *id005
        '500': *id006
      security: *id011
  /api/v1/oracle/ingest/flag:
    post:
      tags:
      - oracle
      summary: Flag project (oracle)
      operationId: oracle_post_flag
      responses:
        '200': *id002
        '400': *id003
        '401': *id004
        '404': *id005
        '500': *id006
      security: *id011
  /api/v1/oracle/price-approvals/hold:
    post:
      tags:
      - oracle
      summary: Hold price approval (admin)
      operationId: oracle_post_hold
      responses:
        '200': *id002
        '400': *id003
        '401': *id004
        '404': *id005
        '500': *id006
      security: *id009
  /api/v1/oracle/price-approvals:
    get:
      tags:
      - oracle
      summary: List price approvals (admin)
      operationId: oracle_get_list
      responses:
        '200': *id002
        '400': *id003
        '401': *id004
        '404': *id005
        '500': *id006
      security: *id009
  /api/v1/oracle/price-approvals/{id}/approve:
    post:
      tags:
      - oracle
      summary: Approve price (admin)
      operationId: oracle_post_approve
      responses:
        '200': *id002
        '400': *id003
        '401': *id004
        '404': *id005
        '500': *id006
      security: *id009
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
  /api/v1/oracle/price-approvals/{id}/reject:
    post:
      tags:
      - oracle
      summary: Reject price (admin)
      operationId: oracle_post_reject
      responses:
        '200': *id002
        '400': *id003
        '401': *id004
        '404': *id005
        '500': *id006
      security: *id009
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
...
  /api/v1/audit:
    get:
      tags:
      - audit
      summary: List audit records (admin)
      operationId: audit_get_list
      responses:
        '200': *id002
        '400': *id003
        '401': *id004
        '404': *id005
        '500': *id006
      security: *id007
  /api/v1/audit/verify:
    get:
      tags:
      - audit
      summary: Verify audit trail (admin)
      operationId: audit_get_verify
      responses:
        '200': *id002
        '400': *id003
        '401': *id004
        '404': *id005
        '500': *id006
      security: *id007
...
  /api/v1/v1/credits/batch/{batchId}:
    get:
      tags:
      - public-api
      summary: 'Public: credit batch'
      operationId: 'public_api_get_public:'
      responses:
        '200': *id002
        '400': *id003
        '401': *id004
        '404': *id005
        '500': *id006
      security: *id011
      parameters:
      - name: batchId
        in: path
        required: true
        schema:
          type: string
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
    X-Api-Key:
      type: apiKey
      in: header
      name: X-Api-Key
  schemas:
...
    CreditBatch:
      type: object
      properties:
        id:
          type: string
          format: uuid
        batchId:
          type: string
        serialNumber:
          type: string
        projectId:
          type: string
        amount:
          type: string
        status:
          type: string
          enum:
          - minted
          - retired
        createdAt:
          type: string
          format: date-time
...
    MintCreditsDto:
      type: object
      properties:
        projectId:
          type: string
        amount:
          type: integer
        vintageYear:
          type: integer
    RetireCreditsDto:
      type: object
      properties:
        batchId:
          type: string
        amount:
          type: integer
        beneficiary:
          type: string
        purpose:
          type: string
...